+++ src/main/java/com/coplaning/dao/PassengerContainer.java
@@ -12,7 +12,7 @@ public class PassengerContainer {
 
 	@PrimaryKey
 	@Persistent(valueStrategy = IdGeneratorStrategy.NATIVE)
-	protected Long id = null;
+	protected Integer id = null;
 
 	@Persistent(defaultFetchGroup = "true")
 	protected Passenger passenger = null;
@@ -31,11 +31,11 @@ public class PassengerContainer {
 
 
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
